+++ lib/plug/conn/status.ex
@@ -9,6 +9,7 @@ defmodule Plug.Conn.Status do
     100 => "Continue",
     101 => "Switching Protocols",
     102 => "Processing",
+    103 => "Early Hints",
     200 => "OK",
     201 => "Created",
     202 => "Accepted",